[REFACTOR]: estrutura de pastas de components
