+++ components/NavBar/index.tsx
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import links from '../lib/links';
+import links from '../../lib/links';
 import Menu from './Menu';
 
 const NavBar = () => {
